adding incomplete render3D example
