+++ render3D/index.html
@@ -31,10 +31,6 @@
             background-color:#fff
         }
     </style>
-    <script src="./three.js-r109/build/three.min.js"></script>
-    <script src="./three.js-r109/examples/js/controls/TrackballControls.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/pako/1.0.11/pako.min.js"></script>
-
     <div id="splash" style="display:flex;flex-direction:column">
         <div style="flex:1 1">
         </div>
@@ -48,6 +44,5 @@
         </div>
     </div>
 
-    <script src="./index.js"></script>
 
 </body>
